requirements.txt
python-telegram-bot[job-queue]==20.4
python-dotenv==1.0.0
schedule==1.2.1 